split main page
